+++ packages/vue-cli-plugin-vue-dash/generator/functions/extendPackage.js
@@ -38,35 +38,49 @@ function extendPackage(api, options) {
 		dependencies: {
 			'@cnamts/design-tokens': DesignTokensVersion,
 			'@cnamts/vue-dot': VueDotVersion,
-			'axios': '^0.21.0',
-			'core-js': '^3.8.0',
+			'axios': '^0.21.1',
+			'core-js': '^3.8.2',
 			'custom-event-polyfill': '^1.0.7',
-			'dayjs': '^1.9.6',
+			'dayjs': '^1.10.3',
 			'languages': '^0.1.3',
+			"vue": "^2.6.12",
 			'vue-input-facade': '^1.3.2',
 			'vue-meta': '^2.4.0',
-			'vuetify': '^2.3.19'
+			"vue-router": "^3.4.9",
+			'vuetify': '^2.4.2'
 		},
 		devDependencies: {
-			'@babel/core': '^7.12.9',
-			'@mdi/js': '^5.8.55',
-			'@types/jest': '^26.0.15',
+			'@babel/core': '^7.12.10',
+			'@mdi/js': '^5.9.55',
+			'@types/jest': '^26.0.20',
 			'@types/webfontloader': '^1.6.32',
-			'@typescript-eslint/eslint-plugin': '^4.8.2',
-			'@typescript-eslint/parser': '^4.8.2',
-			'@vue/eslint-config-standard': '^5.1.2',
+			'@typescript-eslint/eslint-plugin': '^4.13.0',
+			'@typescript-eslint/parser': '^4.13.0',
+			"@vue/cli-plugin-babel": "~4.5.10",
+			"@vue/cli-plugin-eslint": "~4.5.10",
+			"@vue/cli-plugin-router": "~4.5.10",
+			"@vue/cli-plugin-typescript": "~4.5.10",
+			"@vue/cli-plugin-unit-jest": "~4.5.10",
+			"@vue/cli-plugin-vuex": "~4.5.10",
+			"@vue/cli-service": "~4.5.10",
+			'@vue/eslint-config-standard': '^6.0.0',
 			'@vue/eslint-config-typescript': '^7.0.0',
-			'eslint': '^7.14.0',
+			"@vue/test-utils": "^1.1.2",
+			'eslint': '^7.18.0',
 			'eslint-plugin-import': '^2.22.1',
 			'eslint-plugin-node': '^11.1.0',
 			'eslint-plugin-promise': '^4.2.1',
-			'eslint-plugin-vue': '^7.1.0',
+			'eslint-plugin-vue': '^7.4.1',
 			'jest-serializer-vue': '^2.0.2',
+			"sass": "^1.32.4",
+			"sass-loader": "^10.1.1",
+			"typescript": "~4.1.3",
 			'vue-class-component': '^7.2.6',
-			'vue-cli-plugin-vuetify': '^2.0.7',
+			'vue-cli-plugin-vuetify': '^2.0.9',
+			"vue-template-compiler": "^2.6.12",
 			'vuetify-loader': '^1.6.0',
 			'webfontloader': '^1.6.28',
-			'webpack': '^4.44.2'
+			'webpack': '^4.46.0'
 		},
 		resolutions: {
 			'jest': '^26.6.3',
@@ -76,7 +90,7 @@ function extendPackage(api, options) {
 	};
 
 	if (options.i18n) {
-		newPackageProperties.dependencies['vue-i18n'] = '^8.22.2';
+		newPackageProperties.dependencies['vue-i18n'] = '^8.22.4';
 	}
 
 	if (options.formBuilder) {
@@ -91,13 +105,13 @@ function extendPackage(api, options) {
 	}
 
 	if (options.cypress) {
-		newPackageProperties.devDependencies['@vue/cli-plugin-e2e-cypress'] = '~4.5.9';
+		newPackageProperties.devDependencies['@vue/cli-plugin-e2e-cypress'] = '~4.5.10';
 
 		newPackageProperties.scripts = {};
 		newPackageProperties.scripts['test:e2e'] = 'vue-cli-service test:e2e --headless';
 		newPackageProperties.scripts['test:e2e:gui'] = 'vue-cli-service test:e2e';
 
-		newPackageProperties.resolutions['cypress'] = '^6.0.0';
+		newPackageProperties.resolutions['cypress'] = '^6.2.1';
 	}
 
 	api.extendPackage(newPackageProperties);